+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Select component: Make the Select component clearable by providing the placeholder as the native html select tag's `value` attribute if the component's `value` is undefined.
+
 ## 7.0.6
 
 ### Fixed